Add dialog and form content
When the form is filled and submitted the dialog box will close, and all input will be sent to the constructors appropriate fields.
As an aside, the submit button will need to have a prevent default event to prevent it from sending information to a server(which does not exist).