Add flag that checks the read button
Required me to push x in the for loop to the readButton function. This lets us look directly at the nth object
in the array. If x.readStat is 'read' then we set the checkbox to true, else set it to false. This is done
during the creation of the element for the sole purpose of simplicity on my part. No need to run a separate check
just push what is known onto the screen. A second event handler will have to be made to change the read status, and
push the change onto the object in the array