Fix duplicate books from being added from the dialog box
This new for loop will iterate through the array, and compare the userBook.title to that to the same in each element of the array.
If the title matches an alert message pops up informing the user that the book is already in the library, and requests they change it
or specify if it is a remake, adaptation, limited edition, or any other appending to the version already in the library. Append:
It also does this for author and pages, as I realized midway through this commit books may have the same title, but not author or
page count.