Add function to by author and title
The function iterates through a loop that sorts all content either ascending or descending based on which header the user clicks.

Currently the sort by pages has no functionality, and sorting by read status is bugged and being worked on

This is not my code for the most part. The source of this code can be found on W3Schools: https://www.w3schools.com/howto/howto_js_sort_table.asp