Add sorting method to Read header
This operates much the same as the function from w3schools, but instead uses the input status of the checkbox in the cell.
If one is true and the other is false, they switch places. Right now the goal is to simply sort the list copying what was shown
from w3schools, but a future endeavor may be to use something a bit more pragmatic, like selection sort.