Add switch algorithm for resorting object array
This is done to keep the array sorted in line with the table on screen, As every addition, subtraction, or modifacation can cause problems with the
sorted table. This is because any instance of modification causes the display to refresh the table, and bases it off the array. With this
algorithm(Similar to selection sort) this problem should be largely solved, as anytime the table is modified in terms of order, the array is updated
with it. For a full look at the logic, there is addendum text file I wrote out walking myself through the steps.