Move table declarations down in CSS file and stretch table
This was done as it should be closer to the main styling content. As for the descision to stretch the table across the page, it is just testing
ascthetic right now. This can change in the future, but I had an image in my mind on how the table should look and want to try it out.