Fix orientation of arrows vs sorting
This bug was by and large one of the more time consuming ones. The solution was frustratingly simple:
run through each header cell. Provided it is not the cell we have just clicked on, resort the array as it was
when first loaded. Then run through a switch which takes the inner Text of the cell just clicked on and uses
the sortTable function with the appropriate variable.