Add temp styling for table
This all is temporary and is from https://codepen.io/nikhil8krishnan/pen/WvYPvv . The gradient colors will be changed, along with a shadow when
the mouse hovers over a table row. There are some problems with the scrolling and transparency. My recommended current strategy is to place all
list items, not in the thead element, but in a separate table body.