Fix overlapping of table headers over table content
By following the example provided, and placing the table header and table content in two separate div containers, the effect
of overflow while having a transparent table is achieved. It also does not require the table head to have a z index of 1.
Anything passing under the table head, will be obfuscated by the head itself. This fix, however, breaks several things in the
JS code. These include:

Bug: Remove button does not properly work|does not remove the 0th element and when romoving 1st element deletes 0th element.
Bug: Sorting is completely broken. Further, the code implementation for sorting is terrible for larger lists of elements
Bug: Changing read to unread, then deleting the element causes multiple instances of the read checkboxes to become unread/read

The first two bugs are likely due to the rearrangment of elements in the HTML. The second bug is likely a similar case, but also
is poorly optimized. Suggest fixing the second bug, by simply sorting the list directly, then drawing the table back onto screen.