Make dummy content an IIfE to execute when the page loads
This removes the need for a function call at the end of the page. While much of this code could be
refactored into classes, only one constructor was really called throughout the program. Refactoring
the add book functions and extending the booklib array to a constructor that extends Book might
require a massive amount of code surgery that at this moment seems a little excessive outside of
being syntactically pleasing